ALL: Remove use of not thread-safe rand() and srand() and use the internal thread-safe implementation eos::common::getRandom() that provides a similar functionality. Fixes EOS-3319

+++ mgm/GroupBalancer.cc
@@ -156,7 +156,7 @@ GroupBalancer::chooseFidFromGroup(FsGroup* group)
 
   while (validFsIndexes.size() > 0) {
     fs_it = group->begin();
-    rndIndex = common::getRandom((uint64_t)0, validFsIndexes.size() - 1);
+    rndIndex = common::getRandom(0ul, validFsIndexes.size() - 1);
     std::advance(fs_it, validFsIndexes[rndIndex]);
     fsid = *fs_it;
     // Accept only active file systems
